Replace existing TaskParser with ComposedTaskParser and support args
With this change the parser changes to look out for applications
with arguments to be specified in composed task definitions rather
than simple names that represent task definitions.

Rather than two parsers - one for simple definitions and one
for composed definitions - there is now just one parser called
TaskParser that can understand both. The parse result
knows if it is for a simple or composed task.
